+++ design-system/datetime-picker/src/components/__tests__/unit/DatePicker.tsx
@@ -1,8 +1,9 @@
 import React from 'react';
 
+import { convertTokens, legacyParse } from '@date-fns/upgrade/v2';
 import { fireEvent, render } from '@testing-library/react';
 // eslint-disable-next-line no-restricted-imports
-import { format, parse } from 'date-fns';
+import { format } from 'date-fns';
 import { mount, shallow } from 'enzyme';
 
 import Select from '@atlaskit/select';
@@ -18,8 +19,7 @@ test('DatePicker requires no props to be used', () => {
 test('DatePicker, custom formatDisplayLabel', () => {
   const dateValue = new Date('06/08/2018').toUTCString();
   const formatDisplayLabel = (date: string, dateFormat: string) => {
-    const parsed = parse(date);
-    return format(parsed, dateFormat);
+    return format(legacyParse(date), convertTokens(dateFormat));
   };
   const expectedResult = 'June/08';
   const datePickerWrapper = mount<DatePicker>(