+++ editor/editor-common/src/utils/date.ts
@@ -1,5 +1,6 @@
-import differenceInCalendarDays from 'date-fns/difference_in_calendar_days';
-import isBefore from 'date-fns/is_before';
+import { legacyParse } from '@date-fns/upgrade/v2';
+import differenceInCalendarDays from 'date-fns/differenceInCalendarDays';
+import isBefore from 'date-fns/isBefore';
 import { InjectedIntl } from 'react-intl';
 
 enum FORMATS {
@@ -77,8 +78,8 @@ export const timestampToIsoFormat = (timestamp: string | number): string => {
 
 export const isPastDate = (timestamp: string | number): boolean => {
   return isBefore(
-    timestampToIsoFormat(Number(timestamp)),
-    timestampToIsoFormat(Number(todayTimestampInUTC())),
+    legacyParse(timestampToIsoFormat(Number(timestamp))),
+    legacyParse(timestampToIsoFormat(Number(todayTimestampInUTC()))),
   );
 };
 
@@ -88,7 +89,10 @@ export const timestampToTaskContext = (
 ): string => {
   const curDate = new Date(Number(todayTimestampInUTC()));
   const givenDate = new Date(Number(timestamp));
-  const distance = differenceInCalendarDays(givenDate, curDate);
+  const distance = differenceInCalendarDays(
+    legacyParse(givenDate),
+    legacyParse(curDate),
+  );
   const sameYear = givenDate.getUTCFullYear() === curDate.getUTCFullYear();
 
   if (intl && [-1, 0, 1].indexOf(distance) > -1) {