+++ editor/editor-core/src/plugins/date/utils/internal.ts
@@ -1,10 +1,12 @@
 import { formatDateType, dateTypeToDate, dateToDateType } from './formatParse';
 import { DateType } from '../types';
-import addDays from 'date-fns/add_days';
-import addMonths from 'date-fns/add_months';
-import addYears from 'date-fns/add_years';
+import addDays from 'date-fns/addDays';
+import addMonths from 'date-fns/addMonths';
+import addYears from 'date-fns/addYears';
 
 import { DateSegment } from '../types';
+
+import { legacyParse } from '@date-fns/upgrade/v2';
 
 export function padToTwo(number: number) {
   return number <= 99 ? `0${number}`.slice(-2) : `${number}`;
@@ -173,10 +175,10 @@ export function adjustDate(
 
   const newDate: Date =
     activeSegment === 'day'
-      ? addDays(originalDate, adjustment)
+      ? addDays(legacyParse(originalDate), adjustment)
       : activeSegment === 'month'
-      ? addMonths(originalDate, adjustment)
-      : addYears(originalDate, adjustment);
+      ? addMonths(legacyParse(originalDate), adjustment)
+      : addYears(legacyParse(originalDate), adjustment);
 
   return dateToDateType(newDate);
 }