+++ editor/editor-core/src/plugins/hyperlink/ui/HyperlinkAddToolbar/__tests__/HyperlinkAddToolbar.spec.tsx
@@ -74,11 +74,11 @@ const clock = sinon.useFakeTimers(
 );
 jest.unmock('lodash/debounce');
 
-jest.mock('date-fns/difference_in_calendar_days', () => {
+jest.mock('date-fns/differenceInCalendarDays', () => {
   return jest.fn().mockImplementation(() => -5);
 });
 
-jest.mock('date-fns/distance_in_words_to_now', () => ({
+jest.mock('date-fns/formatDistanceToNow', () => ({
   __esModule: true,
   default: () => 'just a minute',
 }));