+++ editor/editor-core/src/ui/LinkSearch/LinkSearchListItem.tsx
@@ -6,12 +6,14 @@ import styled from 'styled-components';
 import { fontSizeSmall } from '@atlaskit/theme';
 import { N20, N800, N100 } from '@atlaskit/theme/colors';
 import { LinkSearchListItemData } from './types';
-import distanceInWords from 'date-fns/distance_in_words';
-import differenceInCalendarDays from 'date-fns/difference_in_calendar_days';
+import formatDistance from 'date-fns/formatDistance';
+import differenceInCalendarDays from 'date-fns/differenceInCalendarDays';
 import format from 'date-fns/format';
 
 import { injectIntl, InjectedIntlProps } from 'react-intl';
 import messages from '../../messages';
+
+import { legacyParse, convertTokens } from '@date-fns/upgrade/v2';
 
 interface ContainerProps {
   selected: boolean;
@@ -139,15 +141,20 @@ class LinkSearchListItem extends React.PureComponent<
         break;
     }
 
-    if (differenceInCalendarDays(timeStamp, Date.now()) < -7) {
+    if (
+      differenceInCalendarDays(
+        legacyParse(timeStamp),
+        legacyParse(Date.now()),
+      ) < -7
+    ) {
       return this.renderWithSpaces(
         pageActionText,
-        format(timeStamp, 'MMMM DD, YYYY'),
+        format(legacyParse(timeStamp), convertTokens('MMMM DD, YYYY')),
       );
     }
     return this.renderWithSpaces(
       pageActionText,
-      distanceInWords(timeStamp, Date.now()),
+      formatDistance(legacyParse(timeStamp), legacyParse(Date.now())),
       intl.formatMessage(messages.timeAgo),
     );
   }