+++ navigation/atlassian-switcher/src/cross-join/providers/__tests__/unit/product-recommendations-fetch-tests.ts
@@ -7,22 +7,22 @@ describe('product-recommendations-fetch-test', () => {
       DIRECT_ACCESS: [
         {
           resourceId:
-            'ari:cloud:jira-software::site/bd98310f-491e-4af3-89a1-a23c6a9db606',
+            'ari:cloud:jira-software::site/11112222-3333-4444-5555-666677778888',
           userAccessLevel: 'EXTERNAL',
           roleAri: 'ari:cloud:jira-software::role/product/member',
           url:
-            'https://recommendations1.jira-dev.com/secure/BrowseProjects.jspa?selectedProjectType=software',
-          displayName: 'recommendations1',
+            'https://example.jira-dev.com/secure/BrowseProjects.jspa?selectedProjectType=software',
+          displayName: 'example',
           avatarUrl:
             'https://site-admin-avatar-cdn.staging.public.atl-paas.net/avatars/240/lightbulb.png',
         },
         {
           resourceId:
-            'ari:cloud:confluence::site/bd98310f-491e-4af3-89a1-a23c6a9db606',
+            'ari:cloud:confluence::site/11112222-3333-4444-5555-666677778888',
           userAccessLevel: 'EXTERNAL',
           roleAri: 'ari:cloud:confluence::role/product/member',
-          url: 'https://recommendations1.jira-dev.com/wiki',
-          displayName: 'recommendations1',
+          url: 'https://example.jira-dev.com/wiki',
+          displayName: 'example',
           avatarUrl:
             'https://site-admin-avatar-cdn.staging.public.atl-paas.net/avatars/240/lightbulb.png',
         },
@@ -34,37 +34,39 @@ describe('product-recommendations-fetch-test', () => {
   const joinableSitesApiResponse = {
     sites: [
       {
-        cloudId: 'bd98310f-491e-4af3-89a1-a23c6a9db606',
-        url: 'https://recommendations1.jira-dev.com',
-        displayName: 'recommendations1',
+        cloudId: '11112222-3333-4444-5555-666677778888',
+        url: 'https://example.jira-dev.com',
+        displayName: 'example',
         relevance: 1000,
         products: {
           'jira-software.ondemand': {
             collaborators: [],
             productUrl:
-              'https://recommendations1.jira-dev.com/secure/BrowseProjects.jspa?selectedProjectType=software',
+              'https://example.jira-dev.com/secure/BrowseProjects.jspa?selectedProjectType=software',
           },
           'confluence.ondemand': {
             collaborators: [],
-            productUrl: 'https://recommendations1.jira-dev.com/wiki',
+            productUrl: 'https://example.jira-dev.com/wiki',
           },
         },
       },
     ],
   };
 
+  const emptyJoinableSiteResponse = { sites: [] };
+
   afterEach(fetchMock.restore);
 
   test('should return error when fetch could not retrieve data', async () => {
     fetchMock.get(
-      '/v1/product-recommendations?capability=DIRECT_ACCESS&product=jira-software&product=jira-servicedesk&product=jira-core&product=confluence',
+      'https://example.com/v1/product-recommendations?capability=DIRECT_ACCESS&product=jira-software&product=jira-servicedesk&product=jira-core&product=confluence',
       400,
     );
 
     // Apparently you have to wrap in a try catch -> https://github.com/facebook/jest/issues/1700
     async function testFetch() {
       try {
-        await fetchProductRecommendationsInternal();
+        await fetchProductRecommendationsInternal('https://example.com');
       } catch (e) {
         throw new Error('testFetchError');
       }
@@ -75,11 +77,33 @@ describe('product-recommendations-fetch-test', () => {
 
   test('should return joinable sites when fetch succeeds', async () => {
     fetchMock.get(
-      '/v1/product-recommendations?capability=DIRECT_ACCESS&product=jira-software&product=jira-servicedesk&product=jira-core&product=confluence',
-      productRecommendationsApiReponse,
+      'https://example.com/v1/product-recommendations?capability=DIRECT_ACCESS&product=jira-software&product=jira-servicedesk&product=jira-core&product=confluence',
+      new Response(JSON.stringify(productRecommendationsApiReponse), {
+        status: 200,
+        statusText: 'ok',
+      }),
     );
 
-    const joinableSites = await fetchProductRecommendationsInternal();
-    expect(joinableSites).toStrictEqual(joinableSitesApiResponse);
+    return fetchProductRecommendationsInternal('https://example.com').then(
+      joinableSitesResponse => {
+        expect(joinableSitesResponse).toStrictEqual(joinableSitesApiResponse);
+      },
+    );
+  });
+
+  test('should return empty joinable sites when fetch returns 400 "Request requires a domain which is not public"', async () => {
+    fetchMock.get(
+      'https://example.com/v1/product-recommendations?capability=DIRECT_ACCESS&product=jira-software&product=jira-servicedesk&product=jira-core&product=confluence',
+      new Response('', {
+        status: 400,
+        statusText: 'Request requires a domain which is not public',
+      }),
+    );
+
+    return fetchProductRecommendationsInternal('https://example.com').then(
+      joinableSitesResponse => {
+        expect(joinableSitesResponse).toEqual(emptyJoinableSiteResponse);
+      },
+    );
   });
 });