+++ people-and-teams/profilecard/src/internal/relative-date.ts
@@ -1,12 +1,17 @@
-import differenceInMonths from 'date-fns/difference_in_months';
-import isThisMonth from 'date-fns/is_this_month';
-import isThisWeek from 'date-fns/is_this_week';
-import isValid from 'date-fns/is_valid';
+import { legacyParse } from '@date-fns/upgrade/v2';
+import differenceInMonths from 'date-fns/differenceInMonths';
+import isThisMonth from 'date-fns/isThisMonth';
+import isThisWeek from 'date-fns/isThisWeek';
+import isValid from 'date-fns/isValid';
 
 import { RelativeDateKeyType } from '../types';
 
 export function isValidDate(date: Date, today: Date = new Date()) {
-  return !!date.getTime && isValid(date) && date.getTime() <= today.getTime();
+  return (
+    !!date.getTime &&
+    isValid(legacyParse(date)) &&
+    date.getTime() <= today.getTime()
+  );
 }
 
 export default function getRelativeDateKey(
@@ -17,11 +22,11 @@ export default function getRelativeDateKey(
     return null;
   }
 
-  if (isThisWeek(date)) {
+  if (isThisWeek(legacyParse(date))) {
     return 'ThisWeek';
   }
 
-  if (isThisMonth(date)) {
+  if (isThisMonth(legacyParse(date))) {
     return 'ThisMonth';
   }
 
@@ -32,7 +37,10 @@ export default function getRelativeDateKey(
     return 'LastMonth';
   }
 
-  const diffInMonths = differenceInMonths(today, date);
+  const diffInMonths = differenceInMonths(
+    legacyParse(today),
+    legacyParse(date),
+  );
   if (diffInMonths < 6) {
     return 'AFewMonths';
   }