+++ design-system/datetime-picker/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/datetime-picker
 
+## 10.0.10
+
+### Patch Changes
+
+- [`70f0701c2e6`](https://bitbucket.org/atlassian/atlassian-frontend/commits/70f0701c2e6) - Upgrade date-fns to 2.17
+
 ## 10.0.9
 
 ### Patch Changes