+++ design-system/logo/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/logo
 
+## 13.1.1
+
+### Patch Changes
+
+- [`70f0701c2e6`](https://bitbucket.org/atlassian/atlassian-frontend/commits/70f0701c2e6) - Upgrade date-fns to 2.17
+
 ## 13.1.0
 
 ### Minor Changes