+++ editor/conversation/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/conversation
 
+## 16.0.20
+
+### Patch Changes
+
+- [`70f0701c2e6`](https://bitbucket.org/atlassian/atlassian-frontend/commits/70f0701c2e6) - Upgrade date-fns to 2.17
+
 ## 16.0.19
 
 ### Patch Changes