+++ elements/date/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/date
 
+## 0.8.5
+
+### Patch Changes
+
+- [`70f0701c2e6`](https://bitbucket.org/atlassian/atlassian-frontend/commits/70f0701c2e6) - Upgrade date-fns to 2.17
+
 ## 0.8.4
 
 ### Patch Changes